port mint integration tests to server v2
