+++ tests/integration/v2/mint/module_test.go
@@ -1,0 +1,51 @@
+package mint
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+
+	"cosmossdk.io/depinject"
+	"cosmossdk.io/log"
+	_ "cosmossdk.io/x/accounts"  // import as blank for app wiring
+	_ "cosmossdk.io/x/bank"      // import as blank for app wiring
+	_ "cosmossdk.io/x/consensus" // import as blank for app wiring
+	_ "cosmossdk.io/x/mint"      // import as blank for app wiring
+	"cosmossdk.io/x/mint/types"
+	_ "cosmossdk.io/x/staking" // import as blank for app wiring
+
+	"github.com/cosmos/cosmos-sdk/tests/integration/v2"
+	"github.com/cosmos/cosmos-sdk/testutil/configurator"
+	_ "github.com/cosmos/cosmos-sdk/x/auth" // import as blank for app wiring
+	authkeeper "github.com/cosmos/cosmos-sdk/x/auth/keeper"
+	_ "github.com/cosmos/cosmos-sdk/x/auth/tx/config" // import as blank for app wiring
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	_ "github.com/cosmos/cosmos-sdk/x/genutil" // import as blank for app wiring
+)
+
+func TestItCreatesModuleAccountOnInitBlock(t *testing.T) {
+	t.Helper()
+	var accountKeeper authkeeper.AccountKeeper
+
+	moduleConfigs := []configurator.ModuleOption{
+		configurator.AccountsModule(),
+		configurator.AuthModule(),
+		configurator.BankModule(),
+		configurator.StakingModule(),
+		configurator.TxModule(),
+		configurator.ValidateModule(),
+		configurator.ConsensusModule(),
+		configurator.GenutilModule(),
+		configurator.MintModule(),
+	}
+
+	startupCfg := integration.DefaultStartUpConfig(t)
+	app, err := integration.NewApp(
+		depinject.Configs(configurator.NewAppV2Config(moduleConfigs...), depinject.Supply(log.NewNopLogger())),
+		startupCfg, &accountKeeper)
+	require.NoError(t, err)
+
+	ctx := app.StateLatestContext(t)
+	acc := accountKeeper.GetAccount(ctx, authtypes.NewModuleAddress(types.ModuleName))
+	require.NotNil(t, acc)
+}
